requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via -r requirements-dev.in
faker==4.0.2              # via -r requirements-dev.in, factory-boy
flake8-commas==2.0.0      # via -r requirements-dev.in
flake8==3.7.9             # via -r requirements-dev.in, flake8-commas
freezegun==0.3.15         # via -r requirements-dev.in
mccabe==0.6.1             # via flake8
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
python-dateutil==2.8.1    # via faker, freezegun
six==1.14.0               # via freezegun, python-dateutil
text-unidecode==1.3       # via faker
